New changes:
Added Pagination.
Refactored code in Books.razor.cs and FilterAndSort.razor
Added new controllers (Genre and Language) in Catalog.API

What needs to be fixed?

Filters: if I clicked on checkbox, I want to see a book, whenever it on this page of catalog or not.

Search bar: if I search books - it will show, whenever this book on the page